Tilemap displayer
So here's the other thing I had done in preperation for the compo: a super basic tiled map displayer, an a camera. I wanted to have a bunch of entity components made so that I could quickly prototype something, but didn't have the time because of messing around with Heaps!

Hoping to optimise this a little bit before the jam starts!

(tileset is a random thing I found on itch for testing out rendering)